[FE] - code refactoring & minor bugfixing
